backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
neo4j>=5.17.0
langgraph>=0.0.26
langchain-openai>=0.0.5
langchain-community>=0.0.19
python-dotenv>=1.0.1
python-multipart>=0.0.9
tavily-python>=0.3.1
